backend/app/requirements.txt
fastapi==0.65
uvicorn==0.14
sqlalchemy==1.3
pydantic[email]==1.8
passlib[bcrypt]==1.7.4
python-jose==3.3
python-multipart==0.0.5
python-dotenv
fastapi-mail
alembic